Started the slideshow component (issue: #7) and also added code to Rover.js
Started a slideshow component that would take in an array of image urls and creates a slideshow from the given image urls. However, need to create the arrows to allow users to move between the different images. In addition, edited the Rover.js file to fetch the images from the json file. However, I have yet to test it to see if it works.

author: Justin Kim